# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MORIZO API — городские квесты
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests/{id}/checkpoints:
    get:
      operationId: listCheckpoints
      tags: [checkpoints]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Checkpoints of a quest
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Checkpoint" }
    post:
      operationId: createCheckpoint
      tags: [checkpoints]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCheckpointBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Checkpoint" }
  /checkpoints/{checkpointId}:
    patch:
      operationId: updateCheckpoint
      tags: [checkpoints]
      parameters:
        - in: path
          name: checkpointId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCheckpointBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Checkpoint" }
    delete:
      operationId: deleteCheckpoint
      tags: [checkpoints]
      parameters:
        - in: path
          name: checkpointId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }

  # ---------- TEAMS ----------
...
  /moderation/quests/{id}/approve:
    post:
      operationId: approveQuest
      tags: [moderation]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }
  /moderation/quests/{id}/reject:
    post:
      operationId: rejectQuest
      tags: [moderation]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RejectQuestBody" }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }

  # ---------- STATS / HOME ----------
...
  /profile:
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update own profile (bio, banner, theme, avatars, city)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /profile/{userId}:
    get:
      operationId: getPublicProfile
      tags: [profile]
      summary: Get public profile of any user with stats
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicProfile" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }

    # ---------- AUTH ----------
...
    User:
      type: object
      required: [id, email, nickname, ageGroup, role, points, createdAt, theme, avatarSlots, activeAvatarSlot]
      properties:
        id: { type: integer }
        email: { type: string }
        nickname: { type: string }
        ageGroup: { type: string, enum: [age_14_15, age_16_17] }
        role: { type: string, enum: [player, moderator] }
        points: { type: integer }
        completedCount: { type: integer }
        bio: { type: string }
        bannerUrl: { type: string }
        city: { type: string }
        theme: { type: string, enum: [neon, sunset, mono] }
        avatarSlots:
          type: array
          items: { $ref: "#/components/schemas/AvatarSlot" }
        activeAvatarSlot: { type: integer }
        createdAt: { type: string, format: date-time }
    AvatarSlot:
      type: object
      required: [style, seed]
      properties:
        style: { type: string }
        seed: { type: string }
    UpdateProfileBody:
      type: object
      properties:
        bio: { type: string, maxLength: 280 }
        bannerUrl: { type: string }
        city: { type: string, maxLength: 64 }
        theme: { type: string, enum: [neon, sunset, mono] }
        avatarSlots:
          type: array
          maxItems: 3
          items: { $ref: "#/components/schemas/AvatarSlot" }
        activeAvatarSlot: { type: integer, minimum: 0, maximum: 2 }
    PublicProfile:
      type: object
      required: [user, achievements, recentSessions]
      properties:
        user: { $ref: "#/components/schemas/User" }
        achievements:
          type: array
          items: { $ref: "#/components/schemas/Achievement" }
        recentSessions:
          type: array
          items: { $ref: "#/components/schemas/Session" }
        rank: { type: integer }
        questsCreated: { type: integer }
...
    Quest:
      type: object
      required: [id, title, description, city, district, difficulty, durationMin, status, authorId, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        city: { type: string }
        district: { type: string }
        coverUrl: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        durationMin: { type: integer }
        rules: { type: string }
        status:
          type: string
          enum: [draft, moderation, published, archived, rejected]
        rejectionReason: { type: string }
        authorId: { type: integer }
        authorNickname: { type: string }
        checkpointCount: { type: integer }
        completionCount: { type: integer }
        avgRating: { type: number }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time }
        bestTravelMode:
          type: string
          enum: [foot, transport, public_transport, dirt_road, off_road]
...
    Checkpoint:
      type: object
      required: [id, questId, orderIndex, name, task, taskType, lat, lng]
      properties:
        id: { type: integer }
        questId: { type: integer }
        orderIndex: { type: integer }
        name: { type: string }
        task: { type: string }
        taskType:
          type: string
          enum: [code_word, choice]
        codeAnswer: { type: string }
        choiceOptions:
          type: array
          items: { type: string }
        choiceAnswerIndex: { type: integer }
        hint: { type: string }
        rules: { type: string }
        lat: { type: number }
        lng: { type: number }
    CreateCheckpointBody:
      type: object
      required: [name, task, taskType, lat, lng]
      properties:
        name: { type: string }
        task: { type: string, minLength: 20 }
        taskType: { type: string, enum: [code_word, choice] }
        codeAnswer: { type: string }
        choiceOptions:
          type: array
          items: { type: string }
        choiceAnswerIndex: { type: integer }
        hint: { type: string }
        rules: { type: string }
        lat: { type: number }
        lng: { type: number }
    UpdateCheckpointBody:
      type: object
      properties:
        name: { type: string }
        task: { type: string }
        taskType: { type: string, enum: [code_word, choice] }
        codeAnswer: { type: string }
        choiceOptions:
          type: array
          items: { type: string }
        choiceAnswerIndex: { type: integer }
        hint: { type: string }
        rules: { type: string }
        lat: { type: number }
        lng: { type: number }
        orderIndex: { type: integer }

    # ---------- TEAMS ----------
...
    Session:
      type: object
      required: [id, questId, mode, status, currentIndex, score, startedAt]
      properties:
        id: { type: integer }
        questId: { type: integer }
        questTitle: { type: string }
        questCity: { type: string }
        questDifficulty: { type: integer }
        teamId: { type: integer }
        teamName: { type: string }
        mode: { type: string, enum: [solo, team] }
        status: { type: string, enum: [started, in_progress, completed, abandoned] }
        travelMode: { type: string, enum: [foot, transport, public_transport, dirt_road, off_road] }
        currentIndex: { type: integer }
        score: { type: integer }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        scoreBreakdown: { $ref: "#/components/schemas/ScoreBreakdown" }
...
    ScoreBreakdown:
      type: object
      required: [checkpointPoints, completionBonus, speedBonus, elapsedSeconds, difficultyMultiplier, travelMultiplier, subtotal]
      properties:
        checkpointPoints: { type: integer }
        completionBonus: { type: integer }
        speedBonus: { type: integer }
        elapsedSeconds: { type: number }
        difficultyMultiplier: { type: number }
        travelMultiplier: { type: number }
        subtotal: { type: integer }
...
    RejectQuestBody:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 5 }

    # ---------- STATS ----------
...
    Achievement:
      type: object
      required: [id, code, name, description, earnedAt]
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        earnedAt: { type: string, format: date-time }

    # ---------- CHAT ----------
...